requirements.txt
streamlit 
groq
nltk
cryptography
python-dotenv
